Auto load last session when id == 0 Move ReviewsDTO to convenience folder
