+++ Communication/Communication.csproj
@@ -96,10 +96,10 @@
     <Compile Include="DataTransfer\Reviews\CommentDataDTO.cs" />
     <Compile Include="DataTransfer\Reviews\CommentInfoDTO.cs" />
     <Compile Include="DataTransfer\Reviews\CustomIncidentDTO.cs" />
-    <Compile Include="DataTransfer\Reviews\MemberPenaltySummaryDTO.cs" />
+    <Compile Include="DataTransfer\Reviews\Convenience\MemberPenaltySummaryDTO.cs" />
     <Compile Include="DataTransfer\Reviews\ReviewPenaltyDTO.cs" />
-    <Compile Include="DataTransfer\Reviews\ReviewsDTO.cs" />
-    <Compile Include="DataTransfer\Reviews\ReviewsPenaltySummaryDTO.cs" />
+    <Compile Include="DataTransfer\Reviews\Convenience\ReviewsDTO.cs" />
+    <Compile Include="DataTransfer\Reviews\Convenience\ReviewsPenaltySummaryDTO.cs" />
     <Compile Include="DataTransfer\Reviews\VoteCategoryDTO.cs" />
     <Compile Include="DataTransfer\Reviews\IncidentReviewDataDTO.cs" />
     <Compile Include="DataTransfer\Reviews\IncidentReviewInfoDTO.cs" />
